python topology
